Lucy in the sky with diamonds (Codewars)
You need to print a shape on the screen using asterisk ("*") characters.

The shape that will be returned from print method resembles a diamond,
where the number provided as input represents the number of *’s printed
on the middle line. The line above and below will be centered and will
have 2 less *’s than the middle line. This reduction by 2 *’s for each
line continues until a line with a single * is printed at the top and
bottom of the figure.

Return null if input is even number (as it is not possible to print
diamond with even number).