Caesar Cipher (Codewars)
The action of a Caesar cipher is to replace each plaintext letter with
a different one a fixed number of places up or down the alphabet.

This program performs a variation of the Caesar shift. The shift
increases by 1 for each character (on each iteration).

If the shift is initially 1, the first character of the message to be
encoded will be shifted by 1, the second character will be shifted by
2, etc...

http://www.codewars.com/kata/5508249a98b3234f420000fb