DirReduction, from CodeWars.
Or how I learned to stop worrying and love the java casts. No, I don’t.

Task:
You have to write a function dirReduc which will take an array of
strings and returns an array of strings with the needless directions
removed (W<->E or S<->N side by side).

http://www.codewars.com/kata/550f22f4d758534c1100025a